raspberry_pi/requirements.txt
paho-mqtt==1.6.1
python-dateutil==2.8.2 